+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-**Version 0.5.0 (unreleased)**
+**Version 0.5.0**
  - [SEMVER_MINOR] Palette entries can now reduced, on by default ([#11](https://github.com/shssoichiro/oxipng/issues/11))
  - Don't report that we are in pretend mode if verbosity is set to none
  - Add cargo bench suite ([#7](https://github.com/shssoichiro/oxipng/issues/7))